Prefix IDs and IDREFs for @id uniqueness
Each distinct, usable 'ography URI is mapped to a prefix (i.e.
'og1') at the beginning of the transform. That prefix is then
prepended to @id on each entry, and @data-tapas-ogref on each
reference. The only exception is 'og0', which always refers to any
entries within the input TEI document itself, and which is left off
of IDs and IDREFs.