Reduce redundancy in LESS files
Variables and the TOC rules were moved to tapasG.less, and the
diplomatic and regularized LESS files import this file, which
in turn imports common.less. Not sure what theory of
organization differentiates tapasG from common. If common is
meant to contain rules that can be applied across viewing
packages, work is needed to separate out the rules specific to
TAPAS-generic.

Styling of <placeName> is commented out until there is a method
for determining whether an anchor exists for a given reference.