Populate popups for internal 'ography entries
When an 'ography entry is present inside the current document, its
structure is the same as those for external, dedicated 'ographies,
which are hidden from view. The JS which creates popup dialogs can
then draw from all available 'ography entries.

As a side effect, the headers are duplicating metadata in an
annoying way.