Normalize generated headers for <person>
Some logic is done on name components within the main <persName>,
so that role names are given preceding commas when necessary, and
when the naming convention is "last name, first name", surnames
are separated from other name parts by a following comma. The
naming convention is done on a <persName>-by-<persName> basis,
since there is no guarantee of document-wide consistency.