Change layout of metadata item class
Instead of including <label>s inside the elements that constitute
metadata items, the copy of the original element is now a sibling
of the generated <label>, inside a <div class="og-metadata-item">.
This allows styling to occur on the element but not the label,
which is especially useful for descendant elements such as
<unclear>, which have distinct and distracting visual markers.

Set up template to allow links where  metadata items are themselves
'ography entries. Fixed bug that introduced unintended whitespace.
Increased whitespace delimiters on metadata items and on context
notes. Streamlined styling of metadata labels in relation to the
field content.