Allow subheadings of 'ography entries
The font-family for headings has been changed to a different serif
font than the one used for other text content. If there is more
than one <head> or <label> on an 'ography entry, all after the
first are put into a <span> with a new class 'heading-sub'. This
class slightly reduces the font-size.

<figure>s are now caught and passed off to 'work' mode when they
are included as part of an 'ography entry.