Move LESS mixins to top; accept <head> before an 'ography sometimes
The LESS files have been somewhat restructured to allow 'ography
entry popups to share some styling with .tapas-generic.

If an 'ography has a <head> immediately before it, and either no
following siblings or an immediately-following <head>, no new
heading is generated. The preceding <head>s are assumed to refer
to the 'ography and only the 'ography. This also means that the
'ography may not be TOC-able.