Keep TOC from 'shivering' when the page scrolls
A Javascript function was toggling the TOC's position from 'static'
to 'fixed' and vice versa, which resulted in the TOC jiggling when
the user scrolled, and sometimes disappearing in a buggy way.

The offending function and its event handler were commented out,
and the LESS/CSS given charge of the TOC 'position' rule.

Some updates were made to TOC's styling. The TOC now: has a border,
doesn't appear so close to the main document, uses a smaller font,
and hides the scrollbar unless it's needed. In addition, the
`div.tapas-generic` wrapper has a height of 100% so it encompasses
its children, and the font-size of the #tapasToolbox has been
changed to use the 'pt' measurement instead of 'px'.