Refactor rendition testing
`@rend` is now tested to determine its encoding style: CSS,
rendition ladder (a la WWO; may not use whitespace to delimit), or
keyword. CSS is output as-is. Rendition ladders are not yet
handled.

If using the keyword system, `@rend` is normalized and tokenized,
and the tokens are matched to pre-set CSS rules.

<q> and <list> now use the 'get-reliable-attributes' template.