Refactor diplomatic mode LESS file
Add parametric mix-ins for communicating interventions. Use
nesting and indentation to improve readability. Make sure that
interventions that depend on attribute values test for the
presence of those attributes.

For the most part, the theory behind the LESS rules has been left
the same, though maintenance of this file should be easier now.