Choose a starter view, with 'diplomatic' as default
The XSLT and Javascript no longer assume that the default view is
'diplomatic'. Changing it in the XSLT via the new parameter
'defaultViewClass' will place that view's class on Generic's
output <div> wrapper. The JS tests for that class and uses
'diplomatic' if it can't find one.

Use a namespace for functions defined in contextualItems.js, which
limits the impact on the global namespace to the variable 'Tapas'
on line 2.

More design tweaks.