Fix identifier mismatch between notes and their anchors
The anchors generated beside in-place notes now use the
'generate-unique-id' template and more robust checking for existing
identifiers.

contextualItems.js now uses a cleaner method of namespacing. Only
those functions which are needed outside the scope of the
self-invoking function are assigned to the TAPAS namespace object.
The private functions are declared as variables: available to the
public functions, but not referenceable through the namespace.