Create generalized templates for references and anchors
When the @ref or @target attributes contain more than one pointer,
only the first is used. Some code was added which would add anchors
to represent subsequent pointers and allow dialog boxes to be
accessed. However, the dependence of ref-generated dialogs on @ref
on an HTML element means that the code is fragile, and so it has
been (hopefully temporarily) commented out.